Added edit widget in a frame for the source
The first approach was to use a QMainWindow so that I could make
the buttons and command line edit window dockable. I could not
make this work.  I think I could do this, but the buttons need to
always be there and the command line input could be placed on the left
dock area and simplify things fairly nicely.