Created a sourceedit class and ebabled font sizing
The SourceEdit class is a derived class of QTextEdit so that the function
keyPressEvent can be overridden to accept key press events in the window.

The font sizing is enabled now from the menu using an inline style sheet
loading for the QApplication.  This replaces the old style sheet so it is OK
for font changes, but it may be necessary to build a more elaborate system if
other items need to be in the stylesheet.  The editor colors previously used
in ebe would not work via stylesheet changes.