Simplified the menus
Used the convenience function version of addAction which incorporates creating
a QAction, connecting it to a slot and setting a shortcut.

At this point most of the slots don't yet exist.  They need to be declared as
slots and coded in the right class to be connected up with little effort.