Started experimenting with doxygen
Doxyfile defines the configuration for doxygen to use when it processes the
files.  The variable INPUTS has the files you wish for doxygen to process.

I started with the CommandLine class which is very simple to learn how to get
reasonable output from doxygen. There are some apparently new commands that I
need to learn about to make the generated docs look better.