Tinkered with widgets; revised settings code
I have worked mostly on the floating point window code.  I have made it look
fairly good by placing 4 QListWidgets side by side with no scrollbar and no
borders on the list widgets.  This sets the stage for using an array of
QListItems to get rapid access to register positions in the lists.

The table in the register window takes up too much vertical space and it was
strange trying to learn how to control the width.

List widgets and text edits are superior when it comes to changing size when
the font size changes.  Even with those I think that eventually I will have
to tweak some sizes when the font size changes.  The pattern for using an hbox
for the list widgets was to make them all the same size by default.  I need
the ones with names to be narrower.  I can figure out how wide to make them
based on the font size.  To be done later.

There is still more to do on the settings code. I have prepared a fair amount
of the start-up code, but haven't yet started reading .ebe.ini.