Added ListWidget to disable wheel scrolling
I added a ListWidget class to use from the dock widgets to provide a
QListWidget with no border, no scrollbar and a function to capture wheel
events and ignore them.  Before it was possible to use the wheel to scroll
in the float window and get names and values misaligned.