Added an addStyleSheet function
The use of qApp->setStyleSheet is too crude.  I wrote a function which
maintains a map of stylesheet pieces.  You call it with a key and a style
option:  addStyleSheet ( "tablegrid", "QTableWidget { blah: blah }");
Any future calls with the same key will replace the stylesheet piece. Each
call to addStyleSheet will use qApp->setStyleSheet to set the concatenation
of the pieces.  Order is not maintained.  Be careful with conflicts.