Connected up to define variable to data window
Marking a variable in the editor and right-clicking allows the user to define
a variable with that name in the data window.  At this point the variable
shows up.  It still needs to send a request to gdb to print the data and set
the value in the data window.