Implemented ebe_unbuffer.cpp to solve buffer issue
ebe_unbuffer defines a single class which consists of essentially only a
constructor.  In this constructor the program calls setbuf to disable
buffering on stdout.  Since the constructor is called for a static object, it
is called before main is called, thus making all writes to stdout unbuffered
without bugging the ebe user.  This change also impacted the pty_reader code.