Added source tab menu, moved tools check
The source tab popup menu allows closing a file in a taba and opening a
file in the same or a new tab.

The tools check in mainwindow.cpp was after starting gdb.  On a computer
with no gdb, ebe would get stuck consuming CPU time with no message.
I moved it to before starting gdb.


Former-commit-id: 63f519f96747f7b45d2c54e59ffaf1a3dff8718a
Former-commit-id: 08e33baec4e41f3ececdbe02aa482794296fe5f9